part5/5_02/broadcaster/requirements.txt
asyncio
nats-py
requests